django_blog/requirements.txt
asgiref==3.9.2
Django==5.2.7
django-taggit==6.1.0
sqlparse==0.5.3
tzdata==2025.2
